# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rais Motors showroom management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: paymentStatus
          schema:
            type: string
            enum: [paid, partial, unpaid]
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaleWithDetails"
    post:
      operationId: createSale
      tags: [sales]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleWithDetails"
  /sales/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: getSale
      tags: [sales]
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleWithDetails"
    delete:
      operationId: deleteSale
      tags: [sales]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /sales/{id}/payments:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    post:
      operationId: addSalePayment
      tags: [sales]
      summary: Record a payment against a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Updated sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleWithDetails"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
...
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchase-orders]
      responses:
        "200":
          description: Purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}/receive:
    post:
      operationId: receivePurchaseOrder
      tags: [purchase-orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    delete:
      operationId: deletePurchaseOrder
      tags: [purchase-orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    BikeInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        brand:
          type: string
        model:
          type: string
        category:
          type: string
        color:
          type: string
        engineCc:
          type: integer
          nullable: true
        purchasePrice:
          type: number
        salePrice:
          type: number
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        sku:
          type: string
        imageUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, brand, model, category, salePrice, stock, sku]
    Bike:
      allOf:
        - $ref: "#/components/schemas/BikeInput"
        - type: object
          properties:
            id:
              type: integer
            createdAt:
              type: string
            updatedAt:
              type: string
          required: [id, createdAt, updatedAt]
    CustomerInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        cnic:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [fullName, phone]
    Customer:
      allOf:
        - $ref: "#/components/schemas/CustomerInput"
        - type: object
          properties:
            id:
              type: integer
            createdAt:
              type: string
          required: [id, createdAt]
...
    SaleInput:
      type: object
      properties:
        customerId:
          type: integer
        bikeId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
        discount:
          type: number
        amountPaid:
          type: number
        paymentMethod:
          type: string
        notes:
          type: string
          nullable: true
      required: [customerId, bikeId, quantity, unitPrice, amountPaid, paymentMethod]
    PaymentInput:
      type: object
      properties:
        amount:
          type: number
          exclusiveMinimum: 0
        paymentMethod:
          type: string
        note:
          type: string
          nullable: true
      required: [amount, paymentMethod]
    SaleWithDetails:
      type: object
      properties:
        id:
          type: integer
        invoiceNo:
          type: string
        customerId:
          type: integer
        bikeId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: number
        totalAmount:
          type: number
        amountPaid:
          type: number
        amountDue:
          type: number
        paymentMethod:
          type: string
        paymentStatus:
          type: string
          enum: [paid, partial, unpaid]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
        bike:
          $ref: "#/components/schemas/Bike"
      required:
        - id
        - invoiceNo
        - customerId
        - bikeId
        - quantity
        - unitPrice
        - discount
        - totalAmount
        - amountPaid
        - amountDue
        - paymentMethod
        - paymentStatus
        - createdAt
        - customer
        - bike
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        expenseDate:
          type: string
        category:
          type: string
        description:
          type: string
        paidTo:
          type: string
          nullable: true
        amount:
          type: number
        bankId:
          type: integer
          nullable: true
        bankName:
          type: string
          nullable: true
        paymentMethod:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, expenseDate, category, description, amount, paymentMethod, createdAt]
    ExpenseInput:
      type: object
      properties:
        expenseDate:
          type: string
        category:
          type: string
        description:
          type: string
        paidTo:
          type: string
          nullable: true
        amount:
          type: number
        bankId:
          type: integer
          nullable: true
        paymentMethod:
          type: string
        notes:
          type: string
          nullable: true
      required: [description, amount]
...
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        orderNo:
          type: string
        supplierName:
          type: string
        bikeId:
          type: integer
        bikeName:
          type: string
          nullable: true
        quantity:
          type: integer
        unitCost:
          type: number
        totalCost:
          type: number
        status:
          type: string
        bankId:
          type: integer
          nullable: true
        bankName:
          type: string
          nullable: true
        orderDate:
          type: string
        receivedDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, orderNo, supplierName, bikeId, quantity, unitCost, totalCost, status, orderDate, createdAt]
    PurchaseOrderInput:
      type: object
      properties:
        supplierName:
          type: string
        bikeId:
          type: integer
        quantity:
          type: integer
        unitCost:
          type: number
        bankId:
          type: integer
          nullable: true
        orderDate:
          type: string
        status:
          type: string
        notes:
          type: string
          nullable: true
      required: [supplierName, bikeId, quantity, unitCost]