# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rais Motors showroom management API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  expenseDate: string;
  category: string;
  description: string;
  paidTo?: string | null;
  amount: number;
  bankId?: number | null;
  bankName?: string | null;
  paymentMethod: string;
  notes?: string | null;
  createdAt: string;
}